refactor: service와 controller의 책임 분리
